refactor: 카드 화면 상단 바 컴포넌트 분리.
